server/requirements-dev.txt
celery-types==0.19.0
djangorestframework-stubs[compatible-mypy]==3.14.2
factory_boy==3.3.0
isort==5.12.0
pretty_errors==1.2.25
pydantic==2.3.0
pytest==7.4.0
pytest-clarity==1.0.1
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-html==3.2.0
pytest-sugar==0.9.7
pyupgrade==3.10.1
tox==4.11.0
types-redis==4.6.0.5
